+++ backend/src/main/java/sg/backend/controller/UserController.java
@@ -22,7 +22,7 @@ import sg.backend.service.EmailService;
 @RestController
 @RequestMapping("/api/user")
 @RequiredArgsConstructor
-public class UserApiController {
+public class UserController {
 
     private final EmailService emailService;
